pyproject.toml
[project]
name = "anhenghealthappai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.117.1",
    "hf-xet>=1.1.10",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.33",
    "openai>=1.109.0",
    "pandas>=2.3.2",
    "pytest>=8.4.2",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "torch>=2.8.0",
    "transformers>=4.56.2",
    "uvicorn>=0.37.0",
]
